requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.9
django-allauth==65.10.0
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-auth-kit==0.3.11
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.10
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
structlog==25.5.0
typing_extensions==4.15.0
uritemplate==4.2.0
